refactor : divide comment apis #48
